Add tests for various values for dateStrings
This test loops through various truthy, falsy and 'date'-ish values
for connection option dateStrings: and runs corresponding typecasting
checks.
